+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v2_0_0/UpgradeSamlSpSessionData.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_1_4;
+package com.liferay.saml.persistence.internal.upgrade.v2_0_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.GetterUtil;
